Add files via upload
Added the screenshot of the result obtained by executing the python program of cards.py .The python program consists of the loop tuning three times that asks the user to input the deck number in which the user sees the selected card in.  After running the code three times and asking the user the input for the deck in which the card is , the program returns the card that the user had selected.